Add files via upload
Here, we are importing the necessary modules and libraries. Flask is the main Flask web framework module, while render_template, request, redirect, and url_for are all functions that we use to handle HTTP requests and render HTML templates. We also import flash to display flash messages to the user. We then import login_required and current_user from flask_login, which we use to require authentication for the bill_payment function. We import FlaskForm from flask_wtf and several form fields and validators from wtforms. Finally, we import stripe, which is the Stripe payment gateway API library.