Add files via upload
This code defines the bill_payment function, which is a view function that handles the /bill-payment route. We use the @app.route decorator to specify the URL route, and the methods parameter to indicate that this function can handle both GET and POST requests.

Inside the function, we create an instance of the BillPaymentForm class and pass it to the render_template function, which renders the bill_payment.html template and passes the form as a parameter.